+++ yandex/data_source_yandex_compute_instance_group.go
@@ -467,6 +467,10 @@ func dataSourceYandexComputeInstanceGroup() *schema.Resource {
 							Type:     schema.TypeInt,
 							Computed: true,
 						},
+						"strategy": {
+							Type:     schema.TypeString,
+							Computed: true,
+						},
 					},
 				},
 			},